Move `parley::style::FontFamily` to `styled_text`
