+++ styled_text/src/lib.rs
@@ -16,8 +16,15 @@
 #![allow(clippy::exhaustive_enums)]
 #![allow(clippy::use_self)]
 
+#[cfg(not(any(feature = "std", feature = "libm")))]
+compile_error!("parley requires either the `std` or `libm` feature to be enabled");
+
+extern crate alloc;
+
 mod attributes;
+mod font_family;
 mod generic;
 
 pub use attributes::{Stretch, Style, Weight};
+pub use font_family::FontFamily;
 pub use generic::GenericFamily;
